+++ extensions/web_interface.py
@@ -1,11 +1,11 @@
 """
 /extensions/supplier_stock_files.py
-Server Monitoring System v8.62.72
+Server Monitoring System v8.62.73
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Web interface
 Система мониторинга серверов
-Версия: 8.62.72
+Версия: 8.62.73
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Веб-интерфейс